services/retrieval/requirements.txt
# Core dependencies
sentence-transformers==2.3.1
qdrant-client>=1.15.1
torch>=2.0.0

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.0
lxml==5.1.0

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.0
numpy>=1.24.0

# Optional (for PDF processing if needed later)
pdfplumber==0.10.0

# Development
pytest==7.4.3
black==23.12.0
